python-hanavector-rag-upload/requirements.txt
Flask==2.3.*
langchain
hdbcli
#sentence-transformers
openai
#langchain-experimental
langchain-community
langchain-openai
python-dotenv
PyPDF2